+++ changelog.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 * Fix `Could not deduce: effs ~ (r : rs)` that may occur when using
   a `Member` contraint (a regression introduced in 0.2.4.0)
+* Add `runNatS` convenience function
 
 ## [0.2.4.1] (November 25, 2016)