+++ changelog.md
@@ -1,17 +1,23 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+* Package renamed to `freer-effects` to distinguish it from original `freer`.
+  [#4](https://github.com/IxpertaSolutions/freer-effects/issues/4)
 * Fix `Could not deduce: effs ~ (r : rs)` that may occur when using
   a `Member` contraint (a regression introduced in 0.2.4.0)
-  [!12](https://gitlab.com/queertypes/freer/merge_requests/12)
+  [freer#12](https://gitlab.com/queertypes/freer/merge_requests/12)
 * Add `runNatS` convenience function
-  [!13](https://gitlab.com/queertypes/freer/merge_requests/13)
+  [freer#13](https://gitlab.com/queertypes/freer/merge_requests/13)
 * Add `evalState` and `execState` convenience functions
-  [!14](https://gitlab.com/queertypes/freer/merge_requests/14)
+  [freer#14](https://gitlab.com/queertypes/freer/merge_requests/14)
 * Data constructors of `Yield`, `CutFalse`, `Fresh`, `State` and `Trace`
   are now exposed in addition to `Exc`, `Reader` and `Writer`
+* Fix [freer#13](https://gitlab.com/queertypes/freer/merge_requests/13)
+* Generalised type signature of `asks`.
+  [#7](https://github.com/IxpertaSolutions/freer-effects/issues/7)
 
 ## [0.2.4.1] (November 25, 2016)